requirements.txt
matplotlib>=3.5
numpy>=1.22
pyserial>=3.5      # needed later for HC-05 RFCOMM
pybluez>=0.23      # simple Bluetooth SPP stack
